backend/requirements.txt
fastapi
uvicorn
tortoise-orm
asyncpg
python-multipart
python-jose[cryptography]
passlib[bcrypt]
